PAGOPA-585 refactoring the project
