+++ docker/Dockerfile.native
@@ -12,7 +12,7 @@
 # Then run the container using:
 #
 # docker run -i --rm -p 8080:8080 quarkus/poc-quarkus
-# docker build -f src/main/docker/Dockerfile.native -t quarkus/poc-quarkus-native:native-aarch64 .
+# docker build -f docker/Dockerfile.native -t quarkus/poc-quarkus-native:native-aarch64 .
 # docker run --rm -p 8080:8080 quarkus/poc-quarkus-native:local
 ###
 FROM registry.access.redhat.com/ubi8/ubi-minimal:8.5
@@ -26,5 +26,6 @@ EXPOSE 8080
 USER 1001
 
 CMD ["./application", "-Dquarkus.http.host=0.0.0.0"]
+RUN true
