+++ docker/run_docker.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+docker-compose up -d --remove-orphans --no-recreate
