--- integration-test/config/properties.json
@@ -1,3 +1,0 @@
-{
-  "reporting_enrollment_host": "http://localhost:8080"
-}
