--- integration-test/package.json
@@ -1,13 +1,0 @@
-{
-  "name": "pagopa-spontaneous-payments",
-  "license": "MIT",
-  "version": "1.0.0",
-  "scripts": {
-    "test": "npx cucumber-js --publish -r step_definitions"
-  },
-  "dependencies": {
-    "@cucumber/cucumber": "^8.4.0",
-    "axios": "^0.27.2",
-    "npx": "^10.2.2"
-  }
-}
