+++ integration-test/src/config/.env.dev
@@ -1,0 +1,1 @@
+REPORTING_ENROLLMENT_HOST=https://api.dev.platform.pagopa.it/shared/poc/reporting-orgs-enrollment/v1
