+++ integration-test/src/config/.env.local
@@ -1,0 +1,3 @@
+REPORTING_ENROLLMENT_HOST=http://localhost:8080
+ORG_ENROLLMENT_SUBSCRIPTION_KEY=
+
