+++ integration-test/src/step_definitions/support/common.js
@@ -1,22 +1,21 @@
 const axios = require("axios");
 const fs = require('fs');
 
-let rawdata = fs.readFileSync('./config/properties.json');
-let properties = JSON.parse(rawdata);
-const reporting_enrollment_host = properties.reporting_enrollment_host;
+const reporting_enrollment_host = process.env.REPORTING_ENROLLMENT_HOST;
+const params = new URLSearchParams([['frameworkType', 'micronaut']]);
 
-function get(url) {
-    return axios.get(reporting_enrollment_host + url)
+function get(url, config) {
+    return axios.get(reporting_enrollment_host + url, config)
          .then(res => {
-             return res;
+            return res;
          })
          .catch(error => {
              return error.response;
          });
 }
 
-function post(url, body) {
-    return axios.post(reporting_enrollment_host + url, body)
+function post(url, config, body) {
+    return axios.post(reporting_enrollment_host + url, body, config)
         .then(res => {
             return res;
         })
@@ -25,8 +24,8 @@ function post(url, body) {
         });
 }
 
-function put(url, body) {
-    return axios.put(reporting_enrollment_host + url, body)
+function put(url, config, body) {
+    return axios.put(reporting_enrollment_host + url, body, config)
         .then(res => {
             return res;
         })
@@ -36,8 +35,8 @@ function put(url, body) {
 }
 
 
-function del(url) {
-    return axios.delete(reporting_enrollment_host + url)
+function del(url, config) {
+    return axios.delete(reporting_enrollment_host + url, config)
         .then(res => {
             return res;
         })
