+++ integration-test/src/step_definitions/support/organizations_steps.js
@@ -13,8 +13,8 @@ let service;
 
 // Given
 
-Given('a random id organization', async function () {
-    idOrg = randomOrg();
+Given('an organization with id {string}', async function (id) {
+    idOrg = id;
     // precondition -> deletion possible dirty data
     await removeOrganization(idOrg);
 });
@@ -27,6 +27,7 @@ Given('a valid enrollment', async function () {
     // save data
     organization = responseToCheck.data;
 });
+
 
 // When
 
@@ -46,10 +47,8 @@ When('the organization {string} enrollment', async function (action) {
 });
 
 
-
-
 // Then
+
 Then('the organization gets the status code {int}', function (status) {
     assert.strictEqual(responseToCheck.status, status);
 });
-
