+++ src/main/resources/application-docker-native-local.properties
@@ -8,5 +8,6 @@ azurite.connection=DefaultEndpointsProtocol=http;AccountName=devstoreaccount1;Ac
 azurite.table=organizations
 
 quarkus.native.additional-build-args=--initialize-at-run-time=com.microsoft.azure
+quarkus.native.builder-image=quay.io/quarkus/ubi-quarkus-mandrel-builder-image:22.2-java11
